Операции удаления, редактирования, изменения вещи вынесены в /item
/user_items теперь отвечает только за получение списка вещей пользователя, что, в принципе, и должен делать
