+++ backend/src/modules/item/item.module.ts
@@ -3,9 +3,13 @@ import { ItemController } from './item.controller';
 import { ItemService } from './item.service';
 import { TypeOrmModule } from '@nestjs/typeorm';
 import { ItemEntity } from '#src/modules/user-items/entity/item.entity';
+import { CategoryEntity } from '#src/modules/categories/entity/category.entity';
+import { PhotoEntity } from '#src/modules/photos/entity/photo.entity';
 
 @Module({
-  imports: [TypeOrmModule.forFeature([ItemEntity])],
+  imports: [
+    TypeOrmModule.forFeature([ItemEntity, CategoryEntity, PhotoEntity]),
+  ],
   controllers: [ItemController],
   providers: [ItemService],
 })
