+++ frontend/src/domains/items/pages/edit-item-page/edit-item-page.tsx
@@ -1,28 +1,50 @@
-import React, { FC, useMemo } from 'react';
+import React, { FC, useEffect, useState } from 'react';
 import { useParams } from 'react-router-dom';
-import { useLoadItemsList } from '#src/js/hooks/use-load-items-list';
-import { itemsListSelector } from '#redux/selectors';
+import { currentItemSelector, userDataSelector } from '#redux/selectors';
 import { useSelector } from 'react-redux';
 import { useLoadCategoriesList } from '#src/js/hooks/use-load-categories-list';
 import { PhotosSlideshow } from '#domains/items/components/photos-slideshow/photos-slideshow';
 import { ItemEditPhotoForm } from '#domains/items/components/item-edit-photo-form/item-edit-photo-form';
 import { ItemEditForm } from '#domains/items/components/item-edit-form/item-edit-form';
+import { Operations } from '#redux/operations/operations';
+import { useAppDispatch } from '#redux/store';
+import { ItemActions } from '#redux/reducers/slices/item-slice';
 
 export const EditItemPage: FC = () => {
   const { itemId } = useParams<{ itemId: string }>();
-  const itemsList = useSelector(itemsListSelector);
-  const item = useMemo(() => itemsList?.find((item) => item.id === parseInt(itemId)), [itemId, itemsList]);
+  const dispatch = useAppDispatch();
+  const item = useSelector(currentItemSelector);
+  const userData = useSelector(userDataSelector);
 
-  useLoadItemsList();
+  const [isUserItem, setIsUserItem] = useState<boolean>(false);
+
+  useEffect(() => {
+    if (item?.user && userData?.username) {
+      setIsUserItem(userData.username === item.user);
+    }
+  }, [item, userData]);
+
+  useEffect(() => {
+    dispatch(Operations.getItem(parseInt(itemId)));
+  }, [itemId, dispatch]);
+
+  useEffect(() => {
+    return () => {
+      dispatch(ItemActions.reset());
+    };
+  }, [dispatch]);
+
   useLoadCategoriesList();
 
-  if (!item) {
+  if (!item)
+    return null;
+
+  if (!isUserItem)
     return (
-      <h4 className={`uk-position-center uk-margin-remove uk-text-muted`}>
-        Такой вещи нет
-      </h4>
+      <div className={`uk-flex uk-flex-center uk-flex-wrap uk-padding-small`}>
+        Это не ваша вещь
+      </div>
     );
-  }
 
   return (
     <div className={`uk-flex uk-flex-center uk-flex-wrap uk-padding-small`}>
