+++ frontend/src/js/redux/reducers/slices/item-slice.ts
@@ -1,7 +1,13 @@
 import { PREState, resetPreState } from '#redux/reducers/util/pre-state';
 import { createSlice } from '@reduxjs/toolkit';
 import { Operations } from '#redux/operations/operations';
-import { onError, onFulfilled, onPending } from '#redux/reducers/util/operation-callback';
+import {
+  onError,
+  onErrorSaveResult,
+  onFulfilled,
+  onPending,
+  onPendingSaveResult,
+} from '#redux/reducers/util/operation-callback';
 import { GetItemOperationResult } from '#redux/operations/slices/get-item-operation';
 
 const initialState: PREState<GetItemOperationResult> = {
@@ -20,7 +26,23 @@ const itemSlice = createSlice({
     builder
       .addCase(Operations.getItem.pending, onPending)
       .addCase(Operations.getItem.rejected, onError)
-      .addCase(Operations.getItem.fulfilled, onFulfilled);
+      .addCase(Operations.getItem.fulfilled, onFulfilled)
+
+      .addCase(Operations.deleteItem.pending, onPendingSaveResult)
+      .addCase(Operations.deleteItem.rejected, onErrorSaveResult)
+      .addCase(Operations.deleteItem.fulfilled, resetPreState)
+
+      .addCase(Operations.createItem.pending, onPendingSaveResult)
+      .addCase(Operations.createItem.rejected, onErrorSaveResult)
+      .addCase(Operations.createItem.fulfilled, onFulfilled)
+
+      .addCase(Operations.editItem.pending, onPendingSaveResult)
+      .addCase(Operations.editItem.rejected, onErrorSaveResult)
+      .addCase(Operations.editItem.fulfilled, onFulfilled)
+
+      .addCase(Operations.setItemPhotos.pending, onPendingSaveResult)
+      .addCase(Operations.setItemPhotos.rejected, onErrorSaveResult)
+      .addCase(Operations.setItemPhotos.fulfilled, onFulfilled);
   },
 });
